Adding more debug info to the flaky test.
The test ConnectionPoolTest.testReuse() fails only on the cloudbees server and there does not seem to be anything wrong in the test as such.
The last run showed that the server was not able to close the connection when the test finished, as it was not even aware of the connection.
These debug logs will tell us if the tests are run in parallel or if there is some mismatch in expectations due to timing delays.