Fix test failure
This test was dependent on the mock server NOT closing the connection on receive but was not setting that state on the server handler.
In some cases, when this state was left uncleared from another test & the connection close from server happens fast enough, the test will fail.

Now, clearing up the state in teardown()