Fix servo HTTP metrics.
The HttpClientListener and HttpServerListener were listening for HttpClient/ServerMetricEvents instead of Client/ServerMetricEvents.
This causes ClassCastExceptions while receiving Client/ServerMetricEvents.

The test did not catch this as the listener in the test was a TCP listener attached to an HttpClient. Since all Http*MetricEvents can be converted to *MetricEvents and not the other way around, the tests were happy.