Fixes #195
RxClient.shutdown() was shutting down the eventloop.

This caused the following scenario impossible to work:

1) Create some clients
2) Shutdown all of them
3) Create another client & try to use it.

In the above scenario, at stage 3, the eventloop is shutdown and hence can not be used by the newly created client.

Removing the eventloop shutdown. Since, the eventloop is always a daemon, shutting it down isn't really necessary.
We will think about graceful JVM shutdown scenarios and hence a need to provide a global shutdown on RxNetty later.