Fixing race conditions in UnicastContentSubject.
UnicastContentSubject has a potential to loose terminal events sent to the subject due to a race condition.
The race condition was when a terminal event arrives between draining all buffered notifications and atomically replacing the observer. Since, the observer, was draining all such buffered notifications, on recieving a notifications, the terminal notifications mentioned above were lost.

This subject was copied from RxJava BufferTillSubscriber which had fixed this race. The intention of not using that subject in RxJava was that it was an internal class. The tax of such bugs is high as compared to rxjava changing the subject contract, hence, I am using rxjava's subject now.