Mark connection as disposed before close.
On unsubscribe, without getting a response, the discard connection flag was not set before close() was called.
This lead to the connection which is supposed to be discarded, return to the pool and potentially used by another request.
This will cause a request to immediately fail since the connection will be closed (as a result of conn. release event) before the request is written.

Now, marking the connection as disposed, when the unsubscribe happens before completion of response.