Race condition in ClientRequestResponseConverter.
For connection pool enabled clients, if the content unsubscription is delayed, it can get into a state where the connection is reused before unsubscription from content. In such a case, the content unsubscription will lead to connection closing, even when the connection is being used for another request.

This also fixes a bug where a content dispose would end up closing the connection. The unsubscription action (closes connection) should not be called if there was no subscription.