Removing client name uniqueness.
The client name is just used for reporting metrics and having a unique name bloats metrics (number of metrics reported).
This change removes the uniqueness and has a common suffix "-no-name-" for all clients with no name explicitly specified.
If a unique name is required, it can be set explicitly on the client.