Unsubscribe from auto release timeout if a subscriber comes along.
If the first item gets emitted very quickly, the timer will basically
always be fired. If the timeout is very long, then all those subscriptions
will land in the old generation and very expensive to pick up.

This changeset unsubscribes immediately if a subscriber comes along, making
it more likely that the subscription is still in the young gen and quicker
to clean up.