UnicastContentSubject unsubscribes eagerly
`UnicastContentSubject` delegates buffering to a `BufferUntilSubscriber` subject. However, it subscribes the downstream subscriber to this subject using `Observable.unsubscribe()`. This causes a `SafeSubscriber` subscribing to the buffer subject which unsubscribes on termination of the upstream source.

The following code reproduces the issue:

```java
        UnicastContentSubject<Long> subject = UnicastContentSubject.createWithoutNoSubscriptionTimeout();

        subject.onNext(1l);
        subject.onCompleted();

        subject.flatMap(aLong -> Observable.never()
                                           .doOnUnsubscribe(() -> System.out.println("unsubscribed")))
               .toBlocking().toFuture().get();
```

In the above code the flatmap returns `Observable.never()` which should not get unsubscribed, unless the eventual subscriber unsubscribes.

Using `unsafeSubscribe()` to subscribe to the `BufferUntilSubscriber` will not eagerly unsubscribe.