Errors in the pipeline post response header delivery were lost.
HttpClient response processing unsubscribes from the TCP connection input subject after the HTTP response headers are delivered.
The errors were only trapped by the `ObservableAdapter` and forwarded to the subscriber of the input subject. In this case, since, the input subject is unsubscribed, the error does not get propagated to the subscriber of the `HttpClient.submit()`.

This change makes `ClientRequestResponseConverter` intercept errors in the pipeline and propagate it to the subscriber of the response content.