Connection Provider
Added a public abstraction of Connection Provider that is used to provide any kind of algorithms to select/create a connection for a TCP/HTTP client.

This abstraction is used internally to provide a pooled connection provider and externally can be used either to provide different connection pooling semantics or a complex load balancer to choose which host+connection to use for any request.