Fixes buffer management issue in ServerSentEventDecoder.
The lastEventType and lastEventId buffers were reused but not retained. This would make the buffers recycled after a generated ServerSentEvent (using these buffers) is released. This leads to all sort of issues, from illegal reference count to wrong dangling pointers.

Now, retaining the buffers for lastEventId and lastEventType once (on channel remove or on receiving a new id/type) whenever they are used to create a ServerSentEvent instance. These buffers are released when the handler is removed.