Adding a facility to emit custom events on EventListener.
Currently protocol specific listeners only provide pre-defined events. In cases, when someone is building a custom protocol over an existing protocol (TCP, HTTP), there would be a need of custom events that can be tapped for metrics, feedback to `ConnectionProvider`s etc.

This change adds 4 variants of `onCustomEvent` methods to `EventListener` (Common class for all listeners) with an ability to emit a custom event with duration and error. The intent of these overloads is that it will allow the actual event class to be a constant (to avoid object allocation overhead) for common usecases of emitting events with duration and error.

RxNetty would not emit any events as custom, it will always be emitted by user code.