Added server latency and total request processing times to HttpClientEventListener.
It is useful for clients to get server side latency per request which is the duration between request write completion and response recieve start.
Added this duration to responseHeaderReceived() callback.

requestProcessingComplete() was not called, fixed it now. It now gives the duration between start of header write and completion of response.