Fixes #395
Backporting 0.5.x for the issue of memory bloat when writing bytes/string from outside eventloop.

Added a handler that converts the `String`/`byte[]` to `ByteBuf` and allowing `String` and `byte[]` to be written directly on to the channel.

This moves the `ByteBuf` allocation from the writing thread to the eventloop.