Adding TCP Server interceptors.
Provides a new abstraction of `TcpServerInterceptorChain` that creates a chain starting with an interceptor and ending on a `ConnectionHandler`.  Any number of interceptors can be added in between. Javadoc on this class explains in detail the abstraction and usage.

Also added are two examples of how these interceptors are used.