Fixes #432
`HttpClient` writes `HttpClientRequest` object on the channel but does not flush the channel after the write as it is expected to be flushed after content is written.
Since, writes do not wake up the selector (flush does), when the requests are written from outside the eventloop, this write (as a task scheduled on the selector's queue) will only be processed the next time selector wakes up. Although, not functionally incorrect, this induces perceivable latencies.

Now, adding a `flush` whenever the write happens from outside the eventloop.