Working around a failed test.
There seems to be a bug in WS client when the WS frames come together with the HTTP upgrade completion headers. The client seems to be discarding the frames as part of the upgrade response.
It seems that on travis, WS frames from server are arriving in the same buffer as the headers.