Fixes #490 (Ordering issues with write)
Ordering issue happens when a write happens from a thread other than the eventloop.

Reported issue was for HTTP, RxNetty adds a LastHttpContent at the end of the content stream to mark the end of chunked content. In this particular case, the LastHttpContent gets written before the content stream completes which causes the illegal state in the HTTP encoder.

The reason for the wrong ordering of writes is as follows:

Netty enqueues any writes that happen from outside the eventloop on to the eventloop task queue.
It writes any write that happens from the eventloop, directly without enqueuing it to the task queue.
In this case, request(n) calls to the content Observable happen on the eventloop, so if the producer(source Observable) is faster than the consumer (writes on channel) then request(n) may immediately produce an item which gets written on the channel as opposed to an item produced on the subscribeOn scheduler that gets enqueued to the task queue. This race results in messing up the order of the writes. In the worst case, if the order happens to be such that the LastHttpContent is written before the content stream completion, the HTTP response is malformed.

The fix is to not have write happening on the eventloop, happen before the writes from outside the eventloop. Thus, if there is any write happening from an `Observable` outside the eventloop. Writes on the eventloop also will be enqueued in the queue. This preserves the ordering.