Fixing test flakiness.
Servers in these examples were using `onBackpressureDrop()` for infinite streams.
This leads to test flakiness as when an element is dropped, the output becomes indeterminate w.r.t order/presence of items which leads to test failures.

Now, instead using `onBackpressureBuffer()` with a limit of 10 as the clients need 10 items.