Fixing test flakiness. (#506)
* Fixing test flakiness.

Servers in these examples were using `onBackpressureDrop()` for infinite streams.
This leads to test flakiness as when an element is dropped, the output becomes indeterminate w.r.t order/presence of items which leads to test failures.

Now, instead using `onBackpressureBuffer()` with a limit of 10 as the clients need 10 items.

* Race condition when a subscriber comes late to listen for new channel and the channel has already unregistered.

When the channel is unregistered, all handlers from the pipeline are removed. Hence, the subscriber event does not reach the handler for the subscriber to be notified.
In this case, the code now simply propagates the closed channel ahead in the pipeline.