Better write backpressure
Today any writes to the channel has a `request(1)` semantic that the subscriber always ever just request 1 item from the source.
This is sub-optimal if the source that is writing isn't in-memory. For out-of-process sources, it is much better to ask for more than 1 item to get batching behavior.

This change introduces the following scheme:

- Have a max buffer size (64 now) per channel.
- This demand is equally split among all active subscribers (for more than one `write()` on the channel)
- All subscribers request up to the maximum limit allowed for them (64/n).
- The subscribers then request more when the pending items go below half the max limit (64/2*n)