Named wire loggers.
Wire Logging available for all clients and servers is used to debug network traffic.
In an application that contains multiple clients (or servers), it will be useful if these logs can be enabled per client/server instead of the current model of enabling logs for all the clients.

This change deprecates the existing `enableWireLogging` methods and add a replacement to also include the name of the logger.