Fixed client connection level events
#### Problem

Since the pipelines for client are setup after the connection is established, callbacks like `connect` are not send to the handlers. This was the reason for missing the connection related events for clients.
`PooledConnection` acquire events were not published because the events were getting published before the event listeners were setup.

#### Modification

This change moves the work done for client on `ChannelHandler.connet()` to `ChannelActivityBufferingHandler` which is added on channel creation.

Now, publishing events for pool acquire are published using the event publisher configured on the channel.