update to netty 4.1.4.Final
There is a regression in netty 4.1.3.Final, see:

http://netty.io/news/2016/07/27/4-0-40-Final-4-1-4-Final.html

The 4.1.4.Final version also includes a change to
QueryStringDecoder that broke some of the UriInfoHolder
tests. In particular the path is now decoded. I opted
to change the tests to match the netty behavior. For
context the netty ticket is: https://github.com/netty/netty/issues/5590